requirements.txt
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.1.0
captum==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==3.1.1
contourpy==1.1.1
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.3.0
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-pygments==0.3.0
kiwisolver==1.4.7
lazy-loader==0.4
lime==0.2.0.1
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prompt-toolkit==3.0.51
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
pyzmq==27.0.0
referencing==0.35.1
requests==2.32.4
rpds-py==0.20.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shap==0.44.1
six==1.17.0
slicer==0.0.7
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.2.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typing-extensions==4.13.0
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.20.2
